feat: 优化生图工具，将图片资源迁移litterFS

+++ src/game/BattleSystem.h
@@ -10,14 +10,14 @@ static constexpr uint8_t SPECIAL_SLOT_NONE = 0xFF;
 
 struct DamageResult {
     uint16_t damage = 0;
-    uint8_t effectiveness = 100; // percent
+    uint16_t effectiveness = 100; // percent
     bool critical = false;
     bool special = false;
     uint8_t specialSlot = SPECIAL_SLOT_NONE;
     bool statusBlocked = false;
 };
 
-uint8_t typeEffectiveness(TypeId attack, TypeId defend1, TypeId defend2);
+uint16_t typeEffectiveness(TypeId attack, TypeId defend1, TypeId defend2);
 uint8_t specialTriggerChance(const Game::MonsterRuntime& attacker);
 bool rollSpecialMove(const Game::MonsterRuntime& attacker);
 uint8_t rollSpecialMoveSlot(const Game::MonsterRuntime& attacker);
